Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325353 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: ae55309fb94e854fb4c73eb1a60a433de1fee7e5

+++ src/core/org/apache/jmeter/testelement/property/IntegerProperty.java
@@ -115,7 +115,7 @@ public class IntegerProperty extends NumberProperty
      */
     public double getDoubleValue()
     {
-        return (double) value;
+        return value;
     }
 
     /**
@@ -123,7 +123,7 @@ public class IntegerProperty extends NumberProperty
      */
     public float getFloatValue()
     {
-        return (float) value;
+        return value;
     }
 
     /**
@@ -139,6 +139,6 @@ public class IntegerProperty extends NumberProperty
      */
     public long getLongValue()
     {
-        return (long) value;
+        return value;
     }
 }
